# swagger.yaml
openapi: 3.0.0
paths:
...
  /wishlistlists:
    get:
      operationId: WishlistsController_findAll
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wishlist'
      tags:
        - wishlistlists              
    post:
      operationId: WishlistsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWishlistDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wishlist'
      tags:
        - wishlistlists
      security:
        - bearer: []
  /wishlistlists/{id}:
    get:
      operationId: WishlistsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wishlist'
      tags:
        - wishlistlists
      security:
        - bearer: []
    patch:
      operationId: WishlistsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWishlistDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wishlist'
      tags:
        - wishlistlists
      security:
        - bearer: []
    delete:
      operationId: WishlistsController_removeOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wishlist'
      tags:
        - wishlistlists
      security:
        - bearer: []
...
info:
  title: КупиПодариДай
  description: API сервиса вишлистов
  version: 1.0.0
  contact: {}
...
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/zlocate/KupiPodariDay/1.0.0
  - url: http://localhost:3000
    description: Локальный сервер
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    WishPartial:
      type: object
      properties:
        id:
          type: number
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        name:
          type: string
          minLength: 1
          maxLength: 250
        link:
          type: string
        image:
          type: string
        price:
          type: number
          minimum: 1
        raised:
          type: number
          minimum: 1
        copied:
          type: number
        description:
          type: string
          minLength: 1
          maxLength: 1024
...
    Wishlist:
      type: object
      properties:
        id:
          type: number
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        name:
          type: string
          minLength: 0
          maxLength: 250
        image:
          type: string
        owner:
          $ref: '#/components/schemas/UserPublicProfileResponseDto'
        items:
          type: array
          items:
            $ref: '#/components/schemas/WishPartial'
      required:
        - id
        - createdAt
        - updatedAt
        - name
        - image
        - owner
        - items
...
    UserPublicProfileResponseDto:
      type: object
      properties:
        id:
          type: number
          example: 5
        username:
          type: string
          minLength: 1
          maxLength: 64
          example: user
        about:
          type: string
          minLength: 1
          maxLength: 200
          example: Пока ничего не рассказал о себе
        avatar:
          type: string
          example: https://i.pravatar.cc/300
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - username
        - about
        - avatar
        - createdAt
        - updatedAt
...
    CreateWishlistDto:
      type: object
      properties:
        name: 
          type: string
          example: "Мой вишлист"
        image:
          type: string
          example: "https://i.pravatar.cc/150?img=3"
        itemsId:
          type: array
          items:
            type: number
          example: [1]
    UpdateWishlistDto:
       type: object
       properties:
        name: 
          type: string
          example: "Мой вишлист"
        image:
          type: string
          example: "https://i.pravatar.cc/150?img=3"
        itemsId:
          type: array
          items:
            type: number
          example: [1]
...